refactored semiperimeter calculation

+++ 0101-computer-programming-1/pe-002/triangle.c
@@ -25,8 +25,8 @@ int main() {
         // Calculate Area using Heron's Formula:
 
         // Get semiperimeter
-        // To prevent integer division truncation, one value of the divisor or dividend must be a float or must be setted to float i.e. "(float)a"
-        s = ((float)a + b + c) / 2;
+        // To prevent integer division truncation, one value of the divisor or dividend must be a float or must be setted to float i.e. "(float)a" or "2.0"
+        s = (a + b + c) / 2.0;
         // Calculate Area
         A = sqrt(s*(s-a)*(s-b)*(s-c));
